Allot Revert + Excel FML Name of Student + Firebase new json file + N…

+++ OptimizeController/Finance/financeController.js
@@ -2730,21 +2730,52 @@ exports.renderFinanceOneBankAccountDestroyQuery = async (req, res) => {
         message: "Their is a bug need to fixed immediatley",
         access: false,
       });
-    const account = await BankAccount.findById({ _id: acid });
-    const finance = await Finance.findById({ _id: `${account?.finance}` });
-    finance.bank_account.pull(account?._id);
-    if (finance?.bank_account_count > 0) {
-      finance.bank_account_count -= 1;
-    }
-    if (account?.finance_bank_upi_qrcode) {
-      await deleteFile(account?.finance_bank_upi_qrcode);
-    }
-    await finance.save();
-    await BankAccount.findByIdAndDelete(acid);
     res.status(200).send({
       message: "Finance / Bank Query Deletion Operation Completed 😁",
       access: true,
     });
+    const account = await BankAccount.findById({ _id: acid });
+    const finance = await Finance.findById({ _id: `${account?.finance}` });
+    let flag = false;
+    const all_receipt = await FeeReceipt.find({
+      finance: finance?._id,
+    })
+      .select("application")
+      .populate({
+        path: "application",
+        select: "applicationDepartment",
+        populate: {
+          path: "applicationDepartment",
+          select: "bank_account",
+        },
+      });
+    if (account?.total_repay > 0) {
+      flag = true;
+    } else {
+      for (let ele of all_receipt) {
+        if (
+          `${ele?.application?.applicationDepartment?.bank_account}` ===
+          `${account?._id}`
+        ) {
+          flag = true;
+        }
+      }
+    }
+
+    if (flag == true) {
+      console.log("Abort Bank Account Deletion", flag);
+    } else {
+      console.log("Bank Account Deletion", flag);
+      finance.bank_account.pull(account?._id);
+      if (finance?.bank_account_count > 0) {
+        finance.bank_account_count -= 1;
+      }
+      if (account?.finance_bank_upi_qrcode) {
+        await deleteFile(account?.finance_bank_upi_qrcode);
+      }
+      await finance.save();
+      await BankAccount.findByIdAndDelete(acid);
+    }
   } catch (e) {
     console.log(e);
   }
@@ -7208,7 +7239,7 @@ exports.renderOneOtherFeesStudentListQuery = async (req, res) => {
     const page = req.query.page ? parseInt(req.query.page) : 1;
     const limit = req.query.limit ? parseInt(req.query.limit) : 10;
     const skip = (page - 1) * limit;
-    const { type } = req?.query;
+    const { type, search } = req?.query;
     if (!ofid)
       return res.status(200).send({
         message: "Their is a bug need to fixed immediately",
@@ -7218,19 +7249,49 @@ exports.renderOneOtherFeesStudentListQuery = async (req, res) => {
     var one_of = await OtherFees.findById({ _id: ofid });
     let list = [...one_of?.paid_students, ...one_of?.remaining_students];
     // one_of?.students
-    var all_student = await Student.find({ _id: { $in: list } })
-      .limit(limit)
-      .skip(skip)
-      .select(
-        "studentFirstName studentMiddleName studentLastName photoId studentProfilePhoto studentGRNO studentROLLNO qviple_student_pay_id other_fees_remain_price other_fees_obj other_fees_paid_price"
-      )
-      .populate({
-        path: "other_fees",
-        populate: {
-          path: "fee_receipt fees",
-          select: "receipt_file fee_payment_amount payable_amount",
-        },
-      });
+    if (search) {
+      var all_student = await Student.find({
+        $and: [{ _id: { $in: list } }],
+        $or: [
+          {
+            studentFirstName: { $regex: `${search}`, $options: "i" },
+          },
+          {
+            studentMiddleName: { $regex: `${search}`, $options: "i" },
+          },
+          {
+            studentLastName: { $regex: `${search}`, $options: "i" },
+          },
+          {
+            studentGRNO: { $regex: `${search}`, $options: "i" },
+          },
+        ],
+      })
+        .select(
+          "studentFirstName studentMiddleName studentLastName photoId studentProfilePhoto studentGRNO studentROLLNO qviple_student_pay_id other_fees_remain_price other_fees_obj other_fees_paid_price"
+        )
+        .populate({
+          path: "other_fees",
+          populate: {
+            path: "fee_receipt fees",
+            select: "receipt_file fee_payment_amount payable_amount",
+          },
+        });
+    } else {
+      var all_student = await Student.find({ _id: { $in: list } })
+        .limit(limit)
+        .skip(skip)
+        .select(
+          "studentFirstName studentMiddleName studentLastName photoId studentProfilePhoto studentGRNO studentROLLNO qviple_student_pay_id other_fees_remain_price other_fees_obj other_fees_paid_price"
+        )
+        .populate({
+          path: "other_fees",
+          populate: {
+            path: "fee_receipt fees",
+            select: "receipt_file fee_payment_amount payable_amount",
+          },
+        });
+    }
     for (let ele of all_student) {
       for (let val of ele?.other_fees) {
         if (`${val?.fees?._id}` === `${one_of?._id}` && val?.fee_receipt) {
@@ -8046,19 +8107,35 @@ exports.renderOneNonExistingOtherFeesStudentListQuery = async (req, res) => {
     const { ofid } = req?.params;
     const page = req.query.page ? parseInt(req.query.page) : 1;
     const limit = req.query.limit ? parseInt(req.query.limit) : 10;
+    const { search } = req?.query;
     if (!ofid)
       return res.status(200).send({
         message: "Their is a bug need to fixed immediately",
         access: false,
       });
 
-    var one_of = await OtherFees.findById({ _id: ofid }).populate({
-      path: "fee_receipt_student",
-      populate: {
-        path: "fee_receipt",
-        select: "receipt_file fee_payment_amount payable_amount",
-      },
-    });
+    if (search) {
+      var one_of = await OtherFees.findById({ _id: ofid }).populate({
+        path: "fee_receipt_student",
+        match: {
+          $or: {
+            student: { $regex: `${search}`, $options: "i" },
+          },
+        },
+        populate: {
+          path: "fee_receipt",
+          select: "receipt_file fee_payment_amount payable_amount",
+        },
+      });
+    } else {
+      var one_of = await OtherFees.findById({ _id: ofid }).populate({
+        path: "fee_receipt_student",
+        populate: {
+          path: "fee_receipt",
+          select: "receipt_file fee_payment_amount payable_amount",
+        },
+      });
+    }
     var all_student = await nested_document_limit(
       page,
       limit,
